keyboard-buttons-bot/requirements.txt
anyio==4.2.0
certifi==2023.11.17
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
hupper==1.12
idna==3.6
python-telegram-bot==20.7
sniffio==1.3.0
python-dotenv==1.0.1